requirements.txt
streamlit>=1.30.0
pandas>=2.1.0
plotly>=5.18.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
pdfplumber>=0.10.0
pyyaml>=6.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
lxml>=4.9.0
anthropic>=0.40.0
